requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
Django==5.2.6
django-cors-headers==4.8.0
django-ninja==1.4.3
pydantic==2.11.7
pydantic_core==2.33.2
sqlparse==0.5.3
typing-inspection==0.4.1
typing_extensions==4.15.0
